requirements.txt
Flask==3.1.2
flask-cors==6.0.1
pyodbc==5.3.0
bcrypt==4.2.1
PyJWT==2.10.1
gunicorn==23.0.0
python-dotenv==1.2.1
requests==2.32.3

# Audio / Video
moviepy==1.0.3
pydub==0.25.1
ffmpeg-python==0.2.0
imageio-ffmpeg==0.6.0
TTS==0.22.0 

# Google Cloud
google-cloud-speech==2.34.0
google-cloud-texttospeech==2.33.0

# OpenAI + AWS
openai==1.65.3
boto3==1.35.13

# LangChain RAG Stack
chromadb==0.5.4
langchain==0.2.14
langchain-community==0.2.11
langchain-text-splitters==0.2.2
langchain-openai==0.1.23  # ✅ Compatible with langchain-core 0.2.x
# ❌ Removed langchain-huggingface to avoid core >= 0.3.0 conflict

# Embeddings (Local)
sentence-transformers==2.2.2
transformers==4.35.2
tokenizers==0.14.1
huggingface-hub==0.17.3   # ✅ Compatible with transformers & tokenizers

# PDF / OCR
pypdf==6.1.3
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.4.0

# Misc
tiktoken==0.11.0
torchcodec
phonemizer
openai-whisper

# Phonetics
rapidfuzz==3.6.1
faster-whisper

numpy
soundfile
eng-to-ipa
torch
torchaudio
torchcodec
